move responsability to compat class
