--- core/src/main/java/com/novoda/accessibility/captioning/CaptionManager.java
@@ -1,7 +1,0 @@
-package com.novoda.accessibility.captioning;
-
-public interface CaptionManager {
-
-    boolean isClosedCaptioningEnabled();
-
-}
