--- core/src/main/java/com/novoda/accessibility/captioning/ClosedCaptionManagerCompat.java
@@ -1,17 +1,0 @@
-package com.novoda.accessibility.captioning;
-
-public class ClosedCaptionManagerCompat implements CaptionManager {
-
-    public static CaptionManager newInstance() {
-        return new ClosedCaptionManagerCompat();
-    }
-
-    public ClosedCaptionManagerCompat() {
-        // no-op
-    }
-
-    @Override
-    public boolean isClosedCaptioningEnabled() {
-        return false;
-    }
-}
